+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.22...HEAD) ##
+## [Unreleased](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.22...HEAD)
 
-## [0.0.22](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.21...v0.0.22) - 2019-04-29 ##
+## Fixed
+
+- The decision tree no longer relies on 'immutable'.
+
+## [0.0.22](https://github.com/craft-ai/react-craft-ai-components/compare/v0.0.21...v0.0.22) - 2019-04-29
 
 ## Fixed